Redo approach to provide default theme for styled components
