--- src/components/common/ThemeProvider/index.ts
@@ -1,1 +1,0 @@
-export * from './ThemeProvider';
